// src/pages/Main/models/ICoin.ts
export interface ICoin {
    id: string;
    name: string;
    image: string;
    current_price: number;
    market_cap: number;
    price_change_percentage_24h: number;
    symbol: string;
    total_volume: number;
    circulating_supply: number;
    high_24h: number;
    atl_date: string;
    total_supply: number;
    low_24h: number;
    fully_diluted_valuation: number;
    max_supply: number;




}